+++ schemars/Cargo.toml
@@ -19,11 +19,12 @@ serde_json = "1.0"
 dyn-clone = "1.0"
 
 chrono = { version = "0.4", default-features = false, optional = true }
-indexmap = { version = "1.2", features=["serde-1"], optional = true }
+indexmap = { version = "1.2", features = ["serde-1"], optional = true }
 either = { version = "1.3", default-features = false, optional = true }
 uuid = { version = "0.8", default-features = false, optional = true }
 smallvec = { version = "1.0", optional = true }
 arrayvec = { version = "0.5", default-features = false, optional = true }
+time = { version = "0.2", optional = true }
 
 [dev-dependencies]
 pretty_assertions = "0.6.1"
